LogCategoryId and LogId now use uint16_t instead of unsigned as underlying type (reduce size type)

+++ src/acl/auth_api.hpp
@@ -24,7 +24,7 @@
 #include "utils/sugar/array_view.hpp"
 
 
-enum class LogId : unsigned;
+enum class LogId : uint16_t;
 
 struct KVLog
 {
